separated fc application
