+++ app/flight_computer_subscale/BSP_L476/CmakeLists.txt
@@ -1,0 +1,13 @@
+add_library(flight_computer_subscale_bsp  STATIC
+    flight_computer_subscale_bsp.c
+)
+target_include_directories(flight_computer_subscale_bsp PUBLIC .)
+target_link_libraries(flight_computer_subscale_bsp
+    PUBLIC
+        app
+        core
+        driver
+        st_l4_support
+    INTERFACE
+        freertos_kernel
+)
