--- app/subscale/BSP_L476/CmakeLists.txt
@@ -1,14 +1,0 @@
-add_library(subscale_bsp STATIC
-    subscale_bsp.c
-)
-
-target_include_directories(subscale_bsp PUBLIC .)
-target_link_libraries(subscale_bsp
-    PUBLIC
-        app
-        core
-        driver
-        st_l4_support
-    INTERFACE
-        freertos_kernel
-)
